+++ lib/opencloud_api_wrapper.dart
@@ -1,1 +1,2 @@
 export 'src/universe.dart' show Universe;
+export 'src/datastore.dart' show Datastore;